accounts: move core-geth multi-coin implementation to separate files
This is intended to help avoid merge conflict issues
in the future.

Unfortunately I have to modify the TestHDPathParsing
test function, adding the SetCoinTypeConfiguration call
to (re-)establish the default values for derivation paths.

This is because
- the configurations are defined as globals,
- and we redefine them for the core-geth implementation

Date: 2022-07-06 14:20:13-07:00
