+++ accounts/hd.go
@@ -25,46 +25,23 @@ import (
 	"strings"
 )
 
-// https://github.com/satoshilabs/slips/blob/master/slip-0044.md
-const BIP0044CoinTypeTestnet uint32 = 0x1
-const BIP0044CoinTypeEther uint32 = 0x3c
-const BIP0044CoinTypeEtherClassic uint32 = 0x3d
-
-// BIP0044CoinType is the global default value for the chain configured hardware derivation path.
-// Its value is set by an init() function and can be modified, along with its dependent global variables
-// to change the default coin type by using the function SetCoinTypeConfiguration.
-var BIP0044CoinType uint32
-
 // DefaultRootDerivationPath is the root path to which custom derivation endpoints
-// are appended. As such, the first account will be at m/44'/61'/0'/0, the second
-// at m/44'/61'/0'/1, etc.
-var DefaultRootDerivationPath DerivationPath
+// are appended. As such, the first account will be at m/44'/60'/0'/0, the second
+// at m/44'/60'/0'/1, etc.
+var DefaultRootDerivationPath = DerivationPath{0x80000000 + 44, 0x80000000 + 60, 0x80000000 + 0, 0}
 
 // DefaultBaseDerivationPath is the base path from which custom derivation endpoints
-// are incremented. As such, the first account will be at m/44'/61'/0'/0/0, the second
-// at m/44'/61'/0'/0/1, etc.
-var DefaultBaseDerivationPath DerivationPath
+// are incremented. As such, the first account will be at m/44'/60'/0'/0/0, the second
+// at m/44'/60'/0'/0/1, etc.
+var DefaultBaseDerivationPath = DerivationPath{0x80000000 + 44, 0x80000000 + 60, 0x80000000 + 0, 0, 0}
 
 // LegacyLedgerBaseDerivationPath is the legacy base path from which custom derivation
-// endpoints are incremented. As such, the first account will be at m/44'/61'/0'/0, the
-// second at m/44'/61'/0'/1, etc.
-var LegacyLedgerBaseDerivationPath DerivationPath
-
-// SetCoinTypeConfiguration sets the global coin type configuration to the given value.
-func SetCoinTypeConfiguration(coinType uint32) {
-	BIP0044CoinType = coinType
-	DefaultRootDerivationPath = DerivationPath{0x80000000 + 44, 0x80000000 + coinType, 0x80000000 + 0, 0}
-	DefaultBaseDerivationPath = DerivationPath{0x80000000 + 44, 0x80000000 + coinType, 0x80000000 + 0, 0, 0}
-	LegacyLedgerBaseDerivationPath = DerivationPath{0x80000000 + 44, 0x80000000 + coinType, 0x80000000 + 0, 0}
-}
-
-// init configures the global coin type and root derivation path for Ethereum mainnet.
-func init() {
-	SetCoinTypeConfiguration(BIP0044CoinTypeEther)
-}
+// endpoints are incremented. As such, the first account will be at m/44'/60'/0'/0, the
+// second at m/44'/60'/0'/1, etc.
+var LegacyLedgerBaseDerivationPath = DerivationPath{0x80000000 + 44, 0x80000000 + 60, 0x80000000 + 0, 0}
 
 // DerivationPath represents the computer friendly version of a hierarchical
-// deterministic wallet account derivation path.
+// deterministic wallet account derivaion path.
 //
 // The BIP-32 spec https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki
 // defines derivation paths to be of the form:
@@ -74,9 +51,9 @@ func init() {
 // The BIP-44 spec https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki
 // defines that the `purpose` be 44' (or 0x8000002C) for crypto currencies, and
 // SLIP-44 https://github.com/satoshilabs/slips/blob/master/slip-0044.md assigns
-// the `coin_type` 61' (or 0x8000003d) to Ethereum.
+// the `coin_type` 60' (or 0x8000003C) to Ethereum.
 //
-// The root path for Ethereum is m/44'/61'/0'/0 according to the specification
+// The root path for Ethereum is m/44'/60'/0'/0 according to the specification
 // from https://github.com/ethereum/EIPs/issues/84, albeit it's not set in stone
 // yet whether accounts should increment the last component or the children of
 // that. We will go with the simpler approach of incrementing the last component.
@@ -175,7 +152,7 @@ func (path *DerivationPath) UnmarshalJSON(b []byte) error {
 }
 
 // DefaultIterator creates a BIP-32 path iterator, which progresses by increasing the last component:
-// i.e. m/44'/61'/0'/0/0, m/44'/61'/0'/0/1, m/44'/61'/0'/0/2, ... m/44'/61'/0'/0/N.
+// i.e. m/44'/60'/0'/0/0, m/44'/60'/0'/0/1, m/44'/60'/0'/0/2, ... m/44'/60'/0'/0/N.
 func DefaultIterator(base DerivationPath) func() DerivationPath {
 	path := make(DerivationPath, len(base))
 	copy(path[:], base[:])
@@ -189,7 +166,7 @@ func DefaultIterator(base DerivationPath) func() DerivationPath {
 
 // LedgerLiveIterator creates a bip44 path iterator for Ledger Live.
 // Ledger Live increments the third component rather than the fifth component
-// i.e. m/44'/61'/0'/0/0, m/44'/61'/1'/0/0, m/44'/61'/2'/0/0, ... m/44'/61'/N'/0/0.
+// i.e. m/44'/60'/0'/0/0, m/44'/60'/1'/0/0, m/44'/60'/2'/0/0, ... m/44'/60'/N'/0/0.
 func LedgerLiveIterator(base DerivationPath) func() DerivationPath {
 	path := make(DerivationPath, len(base))
 	copy(path[:], base[:])
